Support variable features on a symbol "?[var]"
If a feature specifies a string that begins with '?', the remainder is
to be used as a variable name, whose value will be bound in an
environment to the other symbol's value for that feature.

During the rest of the check, if the variable appears again, values
will be compared with those stored in the environment.

Ex: V[sub=?s, obj=?s] == V[sub=sg, obj=sg]
    V[sub=?s, obj=?s] != V[sub=sg, obj=pl]